# docs/openapi.yaml
openapi: 3.1.1
info:
  title: API Lanchonete
  version: 1.0.0
  description: >
    API REST para gerenciamento de pedidos,
    produtos, estoque, usuários, pagamentos
    e unidades da lanchonete.
servers:
  - url: https://lanchonete-api.onrender.com
    description: Servidor de produção

  - url: http://127.0.0.1:5000
    description: Servidor local
...
paths:
...
  /admin/estoque:
    post:
      tags:
      - Registros
      - Estoque
      summary: Adicionar produto ao estoque.
      description: Cadastra produtos no estoque.
      operationId: cadastrarEstoque
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - produtoId
              - unidadeId
              - quantidade
              properties:
                produtoId:
                  type: integer
                  description: ID do produto que será inserido.
                  example: 2
                unidadeId:
                  type: integer
                  description: ID da unidade que receberá o produto.
                  example: 3
                quantidade:
                  type: integer
                  description: Quantidade do produto que a unidade irá receber.
                  example: 50
      responses:
        "201":
          description: Estoque atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "Item adicionado ao estoque com sucesso."
                data:
                  produtoId: 2
                  unidadeid: 2
                  quantidade: 75
                  preco: "3.20"
                  categoria: biscoito
        "400":
          description: Requisição inválida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "JSON_INVALIDO"
                message: "O corpo da requisição deve ser um JSON válido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PRODUTO_NAO_ENCONTRADO"
                message: "O produto informado não foi encontrado."
                details: [{field: "produtoId", issue: "O produto com ID 5 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "409":
          description: Conflito com o estado atual do recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PRODUTO_JA_CADASTRADO"
                message: "O produto informado já se encontra cadastrado."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "422":
          description: Dados inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "QUANTIDADE_INVALIDA"
                message: "A quantidade informada é inválida."
                details: [{field: "quantidade", issue: "A quantidade deve ser maior que zero."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        '401':
          description: Token JWT ausente, expirado ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Usuário autenticado sem permissão para acessar este recurso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /admin/estoque/produtos/{estoqueId}/disponivel/true:
    patch:
      tags:
      - Estoque
      summary: Ativar produto em estoque.
      description: >
        Desativar produto para compra em determinada lunidade. Apenas ADMINISTRADOR, ADMINISTRADOR, GERENCIA e GESTAO têm
        acesso a esta ação.
      operationId: estoqueProdutoAtivo
      security:
      - bearerAuth: []
      parameters:
      - name: estoqueId
        in: path
        required: true
        description: ID do produto no estoque.
        schema:
          type: integer
          minimum: 1
      responses:
        "200":
          description: Produto ativado com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "O produto agora se encontra ativo em estoque."
                data: {idProduto: 2, isActive: false}
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/produtos/2/disponivel/true"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para executar esta ação."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/produtos/2/disponivel/true"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PRODUTO_NAO_ENCONTRADO"
                message: "Não foi possível encontrar este produto."
                details: [{field: "estoqueId", issue: "O produto com o ID 2 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/produtos/2/disponivel/true"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/produtos/2/disponivel/true"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
  /admin/estoque/produtos/{estoqueId}/disponivel/false:
    patch:
      tags:
      - Estoque
      summary: Desativar produto em estoque.
      description: >
        Desativar produto para compra em determinada lunidade. Apenas ADMINISTRADOR, ADMINISTRADOR, GERENCIA e GESTAO têm
        acesso a esta ação.
      operationId: estoqueProdutoInativo
      security:
      - bearerAuth: []
      parameters:
      - name: estoqueId
        in: path
        required: true
        description: ID do produto no estoque.
        schema:
          type: integer
          minimum: 1
      responses:
        "200":
          description: Produto desativado com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "O produto agora se encontra inativo em estoque."
                data: {idProduto: 2, isActive: false}
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/produtos/2/disponivel/false"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para executar esta ação."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/produtos/2/disponivel/false"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PRODUTO_NAO_ENCONTRADO"
                message: "Não foi possível encontrar este produto."
                details: [{field: "estoqueId", issue: "O produto com o ID 2 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/produtos/2/disponivel/false"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/produtos/2/disponivel/false"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
  /admin/estoque/{estoqueId}/quantidade:
    patch:
      tags:
      - Estoque
      summary: Alterar quantidade de itens em estoque.
      description: >
        Permite atualizar a quantidade de um item desejado no pedido. Somente a administração e a gestão, podem realizar esta
        ação.
      operationId: atualizarEstoqueQuantidade
      security:
      - bearerAuth: []
      parameters:
      - name: estoqueId
        in: path
        required: true
        description: Item que terá sua quantidade alterada.
        schema:
          type: integer
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - quantidade
              properties:
                quantidade:
                  type: integer
                  description: Nova quantidade total de itens de determinado produto.
                  example: 5
      responses:
        "200":
          description: Item atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "Quantidade em estoque atualizada com sucesso."
                data:
                  estoqueId: 2
                  unidadeId: 4
                  produtoId: 2
                  quantidade: 60
                  isActive: true
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/2/quantidade"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para realizar esta ação."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/2/quantidade"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PRODUTO_NAO_ENCONTRADO"
                message: "Não foi possível encontrar este produto em estoque."
                details: [{field: "estoqueId", issue: "O produto com o ID 2 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/2/quantidade"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "422":
          description: Dados inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "NOVA_QUANTIDADE_INVALIDA"
                message: "Quantidade informada é menor ou igual a zero."
                details: [{field: "quantidade", issue: "A quantidade deve ser maior que zero."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/2/quantidade"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/admin/estoque/2/quantidade"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
  /unidade/{unidadeId}/menu: ###########
    get:
      tags:
      - Estoque
      summary: Consultar menu da unidade.
      description: Consultar o menu da unidade.
      security:
      - bearerAuth: []
      parameters:
      - name: unidadeId
        in: path
        required: true
        description: ID da unidade cujo menu será consultado.
        schema:
          type: integer
          minimum: 1
      - name: nome
        in: query
        required: false
        description: Filtra pelo nome do produto.
        schema:
          type: string
      - name: categoria
        in: query
        required: false
        description: Filtra pela categoria do produto.
        schema:
          type: string
      - name: disponivel
        in: query
        required: false
        description: Filtra pela disponibilidade do produto.
        schema:
          type: boolean
      - name: valorMin
        in: query
        required: false
        description: Filtra produtos com o menor preço.
        schema:
          type: number
          format: double
          minimum: 0
      - name: valorMax
        in: query
        required: false
        description: Filtra produtos com o maior preço.
        schema:
          type: number
          format: double
          minimum: 0
      - name: ordenar
        in: query
        required: false
        description: Define a ordenação dos resultados.
        schema:
          type: string
          default: nome_asc
          enum:
          - nome_asc
          - nome_desc
          - preco_asc
          - preco_desc
      - name: page
        in: query
        required: false
        description: Número da página.
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: limit
        in: query
        required: false
        description: Quantidade máxima de registros por página.
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      responses:
        "200":
          description: Menu da unidade consultado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuResponse'
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/unidade/2/menu"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para acessar este conteúdo."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/unidade/2/menu"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ESTOQUE_NAO_ENCONTRADO"
                message: "O estoque informado não foi encontrado."
                details: [{field: "estoqueId", issue: "O estoque com ID 2 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/unidade/2/menu"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "422":
          description: Um ou mais filtros informados são inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "FILTRO_INVALIDO"
                message: "O filtro informado é inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/unidade/2/menu"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/unidade/2/menu"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"













      operationId: consultaMenu
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 45
        totalPages:
          type: integer
          example: 3
        hasNext:
          type: boolean
          example: true
        hasPrevious:
          type: boolean
          example: false
    ApiError:
      type: object
      properties:
        error:
          type: string
          example: "USUARIO_NAO_ENCONTRADO"
        message:
          type: string
          example: "O usuário informado não foi encontrado."
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "usuario"
              issue:
                type: string
                example: "O usuário informado não existe."
        timestamp:
          type: string
          format: date-time
          example: "2026-08-07T08:30:00-03:00"
        path:
          type: string
          example: "/login"
        requestId:
          type: string
          format: uuid
          example: "baaf6e1b-5217-41af-9877-2298562e65e3"
...
    RespostaSucesso:
      type: object
      properties:
        message:
          type: string
          example: Operação realizada com sucesso.
        data: {}
...
    MenuResponse:
      type: object
      properties:
        message:
          type: string
          example: Menu da unidade consultado com sucesso.
        data:
          type: array
          items:
            type: object
        meta:
          $ref: '#/components/schemas/PaginationMeta'